+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## 1.1.x
 
+### 1.1.2
+ * Upgrade fount dependnecy to 1.0.0
+ * Allow custom namespace prefix when loading
+
 ### 1.1.1
 Change to allow fount container to be passed in place of full fount instance.